Move donut into it's own component
