+++ app/styles/ui/check-runs/_ci-check-run-popover.scss
@@ -34,7 +34,7 @@
       margin-right: var(--spacing);
 
       .octicon,
-      .donut-chart {
+      .donut {
         display: block;
         margin-top: var(--spacing-half);
         margin-bottom: var(--spacing-half);
@@ -52,7 +52,7 @@
         color: $red-500;
       }
 
-      .donut-chart {
+      .donut {
         > .error,
         > .cancelled,
         > .action_required,
